+++ bot/modules/tools.py
@@ -49,7 +49,10 @@ def is_al_num(string):
 
 
 def date_parser(string):
-    dtx = parser.parse(string, dayfirst=False, tzinfos=TZ_OFFSETS)
+    try:
+        dtx = parser.parse(string, dayfirst=False, tzinfos=TZ_OFFSETS)
+    except parser.ParserError:
+        return
     try:
         dtx = pytz.utc.localize(dtx)
         dtx = dtx.replace(tzinfo=tz.UTC)
@@ -61,3 +64,59 @@ def date_parser(string):
 
 def timestamp_now():
     return int(dt.timestamp(dt.now()))
+
+
+def time_diff(timestamp):
+    lead = timestamp_now() - timestamp
+    if lead < 60:
+        lead_str = f"{lead} second"
+    elif lead < 3600:
+        lead //= 60
+        lead_str = f"{lead} minute"
+    elif lead < 86400:
+        lead //= 3600
+        lead_str = f"{lead} hour"
+    elif lead < 604800:
+        lead //= 86400
+        lead_str = f"{lead} day"
+    elif lead < 2419200:
+        lead //= 604800
+        lead_str = f"{lead} week"
+    else:
+        lead //= 2419200
+        lead_str = f"{lead} month"
+    if lead > 1:
+        return lead_str + "s"
+    else:
+        return lead_str
+
+
+def time_calculator(arg: str):
+    if arg.endswith(('m', 'month', 'months')):
+        time = 2419200
+    elif arg.endswith(('w', 'week', 'weeks')):
+        time = 604800
+    elif arg.endswith(('d', 'day', 'days')):
+        time = 86400
+    elif arg.endswith(('h', 'hour', 'hours')):
+        time = 3600
+    elif arg.endswith(('min', 'mins', 'minute', 'minutes')):
+        time = 60
+    else:
+        return 0
+
+    num = ""
+    for c in arg:
+        if ord('0') <= ord(c) <= ord('9'):
+            num += c
+        else:
+            break
+
+    try:
+        time *= int(num)
+        if time == 0:
+            return 0
+    except ValueError:
+        return 0
+
+    return time